pyproject.toml
[tool.poetry]
name = "finger-detection-server"
version = "0.1.0"
description = "A FastAPI server for detecting fingers in images from ESP32-CAM"
authors = ["Rudr Pratap Singh <rudrpsingh@proton.me>"]

[tool.poetry.dependencies]
python = ">=3.8,<3.14"
fastapi = "^0.103.1"
uvicorn = "^0.23.2"
numpy = "^1.26.0"
opencv-python-headless = "^4.8.1"
python-multipart = "^0.0.6"
pillow = "^10.0.0"
pydantic = "^1.10.8"  # Downgraded to v1 which is more stable with FastAPI

[tool.poetry.dev-dependencies]
pytest = "^7.4.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
